(theme): update colorscheme to improve legibility
I've had a really annoying time configuring dark colors for my terminal.
I'd like it so that on tmux (and some other places such as emacs) I
could reverse the foreground/background colors but there doesn't seem
to be an easy way to do that. So far I just set the ascii color for
black to match the foreground but that made things really hard to read.

Now I'm distinguishing between the foreground color and dark colors on
my theme :) and explicitly setting the background to match the default
foreground wherever I see the need. Its not DRY, but its effective :smile:.