(alpha): Apply on all backgrounds
So the default alpha patch that I applied before only sets the alpha
value for the the bg color, meaning any cells that specify their own
background maintain the default opacity of 1.0. This patch makes it so
that every pixel gets it's background color from either alpha or
alpha_unfocused.

Clearly this is going to have some performance costs, calculations which
we used to do only when we focus the window or at startup we now do for
every cell every time we render. For now I've hidden this new feature
behind a pre-processor macro which you can define to get things working.